// src/api/types.api.ts
export type APIError = {
  reason: string;
};

export type SignUpResponse = {
  id: number;
};

export type UserDTO = {
  id: number;
  login: string;
  first_name: string;
  second_name: string;
  display_name: string;
  avatar: string;
  phone: string;
  email: string;
};

export type CreateUser = Omit<UserDTO, "avatar" | "display_name" | "id"> & {
  password: string;
};

export type CreateChat = {
  title: string;
};

export type LoginRequestData = {
  login: string;
  password: string;
};

export type LastMessage = {
  user: UserDTO;
  time: string;
  content: string;
};

export type ChatDTO = {
  id: number;
  title: string;
  avatar: string | null;
  unread_count: number;
  last_message: LastMessage | null;
};

export type DeleteChat = {
  chatId: number;
};

export type UsersData = {
  users: string[];
  chatId: number;
};

export type ProfileDTO = {
  first_name: string;
  second_name: string;
  display_name: string;
  login: string;
  email: string;
  phone: string;
};

export type PasswordDTO = {
  oldPassword: string;
  newPassword: string;
};

export type Avatar = {
  avatar: File;
};

export type Message = {
  chat_id: number;
  content: string;
  file: File;
  id: number;
  is_read: boolean;
  time: Date | string;
  type: string;
  user_id: string;
};